Add example star wars data
Data is taken from the star wars API project, when the app first starts up the database will be created and populated with the data.

Wouldn't normally use regions, but for something like this with big sections of data and not particularly important code will make an exception.

Update the data model to handle things seen by the data, e.g. wrong types and nullability of things.